Add basic block collection behavior.
If the player collides with a block, it will be freed from the pool (i.e. removed from the game).
If the block is a different color from the player, the player will shake for 30 frames, and the 'PlayerPropulsion' particle system will slow down its particle spawning during that time.